requirements.txt
requests
streamlit
python-docx
PyPDF2